Telagram-Bot-using-GenAI/requirements.txt
python-dotenv
aiogram
langchain-google-genai
google-generativeai
